Dissociate debug and release implementations of LogUtil
