+++ app/src/release/java/com/github/multimatum_team/multimatum/LogUtilImpl.kt
@@ -1,0 +1,16 @@
+package com.github.multimatum_team.multimatum
+
+/**
+ * Release implementation of LogUtil functions (do nothing in release mode)
+ */
+class LogUtilImpl: LogUtil.FunctionsProvider {
+
+    override fun debugLog(str: String) = doNothing()
+
+    override fun logFunctionCall() = doNothing()
+
+    override fun logFunctionCall(str: String) = doNothing()
+
+    private fun doNothing() { }
+
+}
